Fix native commands not using Host UI
Switch to using 'out-host' instead of 'out-default' when running
a command. This forces output from native commands, such as 'ls'
to be sent to the host UI. One problem here is that the output
is formatted differently in some cases. For example 'ls' will
show each item on a separate line when using 'out-host' where as
before it would put them into two columns.